Hash a message from user input
The user can either hash an already existing message or type the message to be hashed. The hash can be saved in a file. If the file already exists, don't make any changes, add to it or overwrite it, based on what the user input is. Read a hash that is either saved in the hashes array or entered by the user. Since SHA256 is irreversible, hash all messages from the messages array until a match is found. If there are no matches, SHA256 can't be reversed.